+++ todo/orm/errors.py
@@ -1,0 +1,2 @@
+class DoesNotExist(Exception):
+    pass
